Changed many uses of "unsigned long" to std::size_t.

+++ src/emacs.cc
@@ -43,7 +43,7 @@ void format_emacs_xacts::write_entry(entry_t& entry)
       break;
     }
 
-  out << (static_cast<unsigned long>(entry.beg_line) + 1) << " ";
+  out << (static_cast<std::size_t>(entry.beg_line) + 1) << " ";
 
   tm when = gregorian::to_tm(*entry.date());
   std::time_t date = std::mktime(&when); // jww (2008-04-20): Is this GMT or local?
@@ -79,7 +79,7 @@ void format_emacs_xacts::operator()(xact_t& xact)
       out << "\n";
     }
 
-    out << "  (" << (static_cast<unsigned long>(xact.beg_line) + 1) << " ";
+    out << "  (" << (static_cast<std::size_t>(xact.beg_line) + 1) << " ";
     out << "\"" << xact.reported_account()->fullname() << "\" \""
 	<< xact.amount << "\"";
